overhaul course cache generation strategy for mutateUserDataUtilServer
* the courseCache in persistUserDataChangesServer() is only needed when
there is a TERM_MOD update chunk in the chunks list. therefore, a new
method, generateCourseCacheFromUpdateChunks(), is created to generate
the course cache from the update chunks (and TERM_MOD update chunk in
particular). We need this instead of generateCourseCacheFlowcharts()
because that doesn't include NEW courses that are being added to a
flowchart.

* also added check to make sure all flowcharts were fetched before
  invoking generateCourseCacheFromUpdateChunks() to make sure we don't
  error out

* instead of throwing errors in generateCourseCacheFromUpdateChunks(),
  we return undefined instead bc these errors can be caused by the user
  passing bad data into the API call. This undefined is detected and
  then causes the persistUserDataChangesServer() call to fail gracefully
  (400 response code to user).