Fix unit counter for standard courses with customUnits (#46)
* adjust computeTermUnits logic to prioritize customUnits

we were previously checking if there was no id on a course before
checking whether we use customUnits. However, this logic is invalid as
there are courses that have an id AND customUnits (e.g. when you edit a
course and have onlyUnitsChange). This updates the logic to reflect this
possibility

* add test to cover standard+customUnits courses