sort fetchedTemplateFlowcharts by order of passed in programIds
order of fetched template flowcharts needs to be in the same order as
programIds provided or else downstream routines will fail. This
indeterminism is what caused some program combinations to error when
creating a new flowchart, as the mergeFlowchartsCourseData() routine
relies on the fetched template flowcharts and data.programIds to have
the programs in the same order (and this was not the case).