add playwright and vitest runner modes for verifyCourseCacheStrictEquality
need to do this bc we want to use this assertion in both playwright and
vitest test runners. if we try to statically import 'vitest' module in
playwright environment, test runner will error out. so, we need to
dynamically import the vitest module ONLY when we're using the vitest
test runner.